requirements-santteegt.txt
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Authlib==1.3.1
av==12.3.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.3
beartype==0.15.0
beautifulsoup4==4.12.3
bidict==0.23.1
binaryornot==0.4.4
black==24.4.2
bleach==6.1.0
blinker==1.8.2
bokeh==3.4.3
boto3==1.34.110
botocore==1.34.149
Brotli==1.1.0
build==1.2.1
cached-property==1.5.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
colorclass==2.2.2
coloredlogs==15.0.1
comm==0.2.2
compressed_rtf==1.0.6
contourpy==1.2.1
cookiecutter==2.6.0
cryptography==43.0.0
ctranslate2==4.3.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docx2txt==0.8
duckduckgo_search==6.1.12
easygui==0.98.3
ebcdic==1.1.1
ecdsa==0.19.0
email_validator==2.2.0
emoji==2.12.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
extract-msg==0.48.7
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.4
faster-whisper==1.0.2
fastjsonschema==2.20.0
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.1
fpdf2==2.7.9
fqdn==1.5.1
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.6.1
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.22.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
h2==4.1.0
h3==3.7.7
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_bokeh==4.0.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kubernetes==30.1.0
langchain==0.2.0
langchain-chroma==0.1.1
langchain-community==0.2.0
langchain-core==0.2.24
langchain-openai==0.1.19
langchain-text-splitters==0.2.2
langdetect==1.0.9
langfuse==2.33.0
langsmith==0.1.93
lark==1.1.9
linkify-it-py==2.0.3
litellm==1.42.5
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msoffcrypto-tool==5.4.1
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
oletools==0.60.2
onnxruntime==1.16.3
open-webui==0.3.10
openai==1.37.1
openapi-pydantic==0.4.1
opencv-python==4.10.0.84
opencv-python-headless==4.9.0.80
openparse==0.5.6
openpyxl==3.1.2
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
ordered-set==4.1.0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
panel==1.4.4
param==2.1.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pathway==0.14.0
pcodedmp==1.2.6
pdf2image==1.17.0
pdfminer.six==20240706
peewee==3.17.5
peewee-migrate==1.12.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pillow_heif==0.18.0
pip==24.2
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
posthog==3.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyMySQL==1.1.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreqwest_impersonate==0.5.3
PySocks==1.7.1
python-dateutil==2.9.0
python-docx==1.1.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-iso639==2024.4.27
python-jose==3.3.0
python-json-logger==2.0.7
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.9
python-sat==1.8.dev13
python-slugify==8.0.4
python-socketio==5.11.2
pytube==15.0.0
pytz==2024.1
pyviz_comms==3.0.2
pyxlsb==1.0.10
PyYAML==6.0.1
pyzmq==26.0.3
rank-bm25==0.2.2
rapidfuzz==3.9.4
rapidocr-onnxruntime==1.3.22
red-black-tree-mod==1.20
redis==5.0.7
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
RTFDE==0.1.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
Send2Trash==1.8.3
sentence-transformers==2.7.0
setuptools==71.1.0
shapely==2.0.5
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spotipy==2.24.0
SQLAlchemy==2.0.31
sqlglot==10.6.1
stack-data==0.6.2
starlette==0.37.2
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.2
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-utils==0.1.0
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
ujson==5.10.0
unstructured==0.14.0
unstructured-client==0.24.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.22.0
uvloop==0.19.0
validators==0.28.1
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
xyzservices==2024.6.0
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.19.2
zstandard==0.22.0
